refactor RestClient as interface
